oko_attached: extract OAuth config and utilities to shared module
